requirements.txt
addict==2.4.0
aenum==3.1.15
affine==2.4.0
click-plugins==1.1.1
cligj==0.7.2
colorlog==6.9.0
dataclasses-json==0.6.7
einops==0.7.0
filelock==3.13.1
h5py==3.10.0
iopath==0.1.10
lightgbm==4.6.0
marshmallow==3.22.0
matplotlib==3.7.5
ml-collections==0.1.1
mypy-extensions==1.0.0
myutils==0.0.21
natsort==8.4.0
opencv-python-headless==4.11.0.86
opencv-python==4.9.0.80
pancollection==0.3.6
portalocker==3.0.0
pyproj==3.5.0
pywavelets==1.4.1
pyyaml==6.0.1
rasterio==1.3.11
regex==2024.11.6
safetensors==0.4.5
sentinelhub==3.11.1
sentry-sdk==2.18.0
setproctitle==1.3.3
shapely==2.0.7
shortuuid==1.0.13
smmap==5.0.1
snuggs==1.4.7
tabulate==0.9.0
tensorboardx==2.6.2.2
termcolor==2.4.0
tiffile==2018.10.18
timm==0.9.11
tokenizers==0.20.3
tomli-w==1.0.0
torchsummary==1.5.1
tqdm==4.66.2
transformers==4.46.2
typing-inspect==0.9.0
udl-vis==0.3.7
utm==0.8.1
yacs==0.1.8
yapf==0.43.0